moved save functions to readsave, reordered deps
