+++ src/ParallelTemperingMonteCarlo.jl
@@ -15,9 +15,10 @@ include("InputParams.jl")
 #include("Initialization")
 include("MCMoves.jl")
 include("Exchange.jl")
+include("ReadSave.jl")
 include("MCRun.jl")
 include("multihist.jl")
-#include("ReadSave.jl")
+
 
 include("parallelrun.jl")
 
@@ -31,10 +32,11 @@ include("parallelrun.jl")
  #@reexport using .Initialization
  @reexport using .MCMoves
  @reexport using .Exchange
+ @reexport using .ReadSave
  @reexport using .MCRun
  @reexport using .Multihistogram
 
- #@reexport using .ReadSave
+ 
 
  @reexport using .ParallelRun
